toolkit/components/satchel/test/test_popup_direction.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Popup Direction</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="satchel_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Test for Popup Direction
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">
  <!-- normal, basic form -->
  <form id="form1" onsubmit="return false;">
    <input  type="text" name="field1">
    <button type="submit">Submit</button>
  </form>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">
var resolvePopupShownListener;
registerPopupShownListener(() => resolvePopupShownListener());

function waitForNextPopup() {
  return new Promise(resolve => { resolvePopupShownListener = resolve; });
}

add_task(async function test_popup_direction() {
  let input = $_(1, "field1");

  await new Promise(resolve => updateFormHistory([
    { op: "remove" },
    { op: "add", fieldname: "field1", value: "value1" },
    { op: "add", fieldname: "field1", value: "value2" },
  ], resolve));

  for (let direction of ["ltr", "rtl"]) {
    document.getElementById("content").style.direction = direction;

    let popupShown = waitForNextPopup();
    input.focus();
    synthesizeKey("KEY_ArrowDown");
    await popupShown;

    let popupState = await new Promise(resolve => getPopupState(resolve));
    is(popupState.direction, direction, "Direction should match.");

    // Close the popup.
    input.blur();
  }
});

</script>
</pre>
</body>
</html>
